Move toString into utils.
