+++ modules/types/src/dummy_type.cpp
@@ -11,7 +11,7 @@
 #include <magic_enum.hpp>
 
 #include "hephaestus/random/random_object_creator.h"
-#include "hephaestus/types/type_formatting.h"
+#include "hephaestus/utils/string/type_formatting.h"
 
 namespace heph::types {
 
@@ -54,12 +54,11 @@ auto DummyType::random(std::mt19937_64& mt) -> DummyType {
 }
 
 auto operator<<(std::ostream& os, const DummyType& dummy_type) -> std::ostream& {
-  return os << "DummyType{"
-            << "\n"
+  return os << "DummyType{\n"
             << "  dummy_primitives_type={" << dummy_type.dummy_primitives_type << "}\n"
             << "  dummy_enum=" << magic_enum::enum_name(dummy_type.dummy_enum) << "\n"
             << "  dummy_string=" << dummy_type.dummy_string << "\n"
-            << "  dummy_vector=" << toString(dummy_type.dummy_vector) << "\n"
+            << "  dummy_vector=" << utils::string::toString(dummy_type.dummy_vector) << "\n"
             << "}";
 }
