+++ modules/utils/src/string/type_formatting.cpp
@@ -2,4 +2,4 @@
 // Copyright (C) 2023-2024 HEPHAESTUS Contributors
 //=================================================================================================
 
-#include "hephaestus/types/type_formatting.h"  // NOLINT(misc-include-cleaner)
+#include "hephaestus/utils/string/type_formatting.h"  // NOLINT(misc-include-cleaner)
